backend/requirements.txt
fastapi==0.129.0
uvicorn==0.40.0
supabase==2.28.0
pydantic==2.12.5
pydantic-settings==2.12.0
python-dotenv==1.2.1
PyJWT[crypto]==2.11.0
python-multipart==0.0.22
httpx==0.28.1
pytest==9.0.2
pytest-asyncio==1.3.0
